requirements.txt
pyautogui==0.9.54

Pillow==9.5.0
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pdf2image==1.16.3
streamlit==1.25.0

augraphy==8.2.4